+++ include/i18n.h
@@ -12,7 +12,10 @@ typedef enum {
     I18N_NORMAL_LATEST,
     I18N_NORMAL_NEW_MESSAGES,
     I18N_HELP_TITLE,
-    I18N_HELP_STATUS_FORMAT
+    I18N_HELP_STATUS_FORMAT,
+    I18N_COMMAND_OUTPUT_TITLE,
+    I18N_MOTD_TITLE,
+    I18N_MOTD_CONTINUE_HINT
 } i18n_text_id_t;
 
 bool i18n_try_parse_lang(const char *value, help_lang_t *lang);